requirements.txt
asgiref==3.8.1
Django==5.1.6
sqlparse==0.5.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
channels==4.2.0
daphne==4.1.2
